+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Sosh theme (Orange-OpenSource/ouds-ios#665)
 - Versions of components in package
 - New content-on color semantic tokens (tokens library v1.1.0) (Orange-OpenSource/ouds-ios#664)
 - New repository color semantic tokens (tokens library v1.1.0) (Orange-OpenSource/ouds-ios#664)